More MinGW64 fixes
Some more trivial fixes for successfully compiling for Win64.
(cherry picked from commit 7b85371a3804b5c33ddd6841d5122f6de12dacdc)

+++ plugins/zynaddsubfx/fltk/src/fl_ask.cxx
@@ -67,7 +67,7 @@ static char avoidRecursion = 0;
 
 // sets the global return value (ret_val) and closes the window
 static void button_cb(Fl_Widget *bt, void *val) {
-  ret_val = (int)(long)val;
+  ret_val = (int)(intptr_t)val;
   message_form->hide();
 }
